+++ radiant/brush/Brush.h
@@ -13,15 +13,19 @@ class RenderableCollector;
 
 const std::size_t c_brush_maxFaces = 1024;
 
-inline double quantiseFloating(double f) {
+inline float quantiseFloating(float f)
+{
 	return float_snapped(f, 1.f / (1 << 16));
 }
 
 /// \brief Returns true if 'self' takes priority when building brush b-rep.
-inline bool plane3_inside(const Plane3& self, const Plane3& other) {
-	if (vector3_equal_epsilon(self.normal(), other.normal(), 0.001f)) {
+inline bool plane3_inside(const Plane3& self, const Plane3& other)
+{
+	if (vector3_equal_epsilon(self.normal(), other.normal(), 0.001f))
+	{
 		return self.dist() < other.dist();
 	}
+
 	return true;
 }
