+++ radiant/brush/Winding.cpp
@@ -179,7 +179,7 @@ BrushSplitType Winding::classifyPlane(const Plane3& plane) const
 	return split;
 }
 
-PlaneClassification Winding::classifyDistance(const double distance, const double epsilon)
+PlaneClassification Winding::classifyDistance(const float distance, const float epsilon)
 {
 	if (distance > epsilon) {
 		return ePlaneFront;
@@ -215,7 +215,7 @@ std::size_t Winding::opposite(const std::size_t index, const std::size_t other) 
 {
 	ASSERT_MESSAGE(index < size() && other < size(), "Winding::opposite: index out of range");
 
-	double dist_best = 0;
+	float dist_best = 0;
 	std::size_t index_best = c_brush_maxFaces;
 
 	Ray edge(ray_for_points((*this)[index].vertex, (*this)[other].vertex));
@@ -226,7 +226,7 @@ std::size_t Winding::opposite(const std::size_t index, const std::size_t other) 
 			continue;
 		}
 
-		double dist_squared = ray_squared_distance_to_point(edge, (*this)[i].vertex);
+		float dist_squared = ray_squared_distance_to_point(edge, (*this)[i].vertex);
 
 		if (dist_squared > dist_best) {
 			dist_best = dist_squared;
@@ -246,13 +246,13 @@ Vector3 Winding::centroid(const Plane3& plane) const
 {
 	Vector3 centroid(0,0,0);
 
-	double area2 = 0, x_sum = 0, y_sum = 0;
+	float area2 = 0, x_sum = 0, y_sum = 0;
 	const ProjectionAxis axis = projectionaxis_for_normal(plane.normal());
 	const indexremap_t remap = indexremap_for_projectionaxis(axis);
 
 	for (std::size_t i = size() - 1, j = 0; j < size(); i = j, ++j)
 	{
-		const double ai = (*this)[i].vertex[remap.x]
+		const float ai = (*this)[i].vertex[remap.x]
 				* (*this)[j].vertex[remap.y] - (*this)[j].vertex[remap.x]
 				* (*this)[i].vertex[remap.y];
